break up logic into new files/hooks
